Resolving this comment made by Steven for the Zen, Zec and BCH forwarders
The payment forward input payment should be created even if the output payment fails. That way we can understand which payments have not yet been forwarded.

In this case the payment forward input payment is only recorded in the database once coins are sent to the output address. forwards.createPaymentForwardInputPayment could be more effective if moved up.